+++ test/test_completion_engine.py
@@ -1,4 +1,5 @@
 from mycli.packages.completion_engine import suggest_type
+import os
 import pytest
 
 
@@ -495,3 +496,18 @@ def test_cross_join():
 def test_after_as(expression):
     suggestions = suggest_type(expression, expression)
     assert set(suggestions) == set()
+
+
+@pytest.mark.parametrize('expression', [
+    '\\. ',
+    'select 1; \\. ',
+    'select 1;\\. ',
+    'select 1 ; \\. ',
+    'source ',
+    'truncate table test; source ',
+    'truncate table test ; source ',
+    'truncate table test;source ',
+])
+def test_source_is_file(expression):
+    suggestions = suggest_type(expression, expression)
+    assert suggestions == [{'type': 'file_name'}]